# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for sports field booking system - Take Home Test",
        "title": "Sports Field Booking API",
        "contact": {
            "name": "API Support",
            "email": "support@sagaratech.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3005",
    "basePath": "/api",
    "paths": {
...
        "/fields": {
            "get": {
                "description": "Get list of all available sports fields. PUBLIC ACCESS - No authentication required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fields"
                ],
                "summary": "Get all fields",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/take-home-test_app_models.FieldResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new sports field. ADMIN ACCESS ONLY - Regular users cannot create fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fields"
                ],
                "summary": "Create new field",
                "parameters": [
                    {
                        "description": "Field data including name, price_per_hour, and location",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.CreateFieldRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/take-home-test_app_models.FieldResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            }
        },
        "/fields/{id}": {
            "get": {
                "description": "Get details of a specific sports field. PUBLIC ACCESS - No authentication required.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fields"
                ],
                "summary": "Get field by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/take-home-test_app_models.FieldResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update sports field information. ADMIN ACCESS ONLY - Regular users cannot update fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fields"
                ],
                "summary": "Update field",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Field update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.UpdateFieldRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/take-home-test_app_models.FieldResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a sports field. ADMIN ACCESS ONLY - Regular users cannot delete fields.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fields"
                ],
                "summary": "Delete field",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Field ID (UUID format)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/take-home-test_app_models.BasicResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "take-home-test_app_models.BasicResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {},
                "pagination": {
                    "$ref": "#/definitions/take-home-test_app_models.Pagination"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "take-home-test_app_models.CreateFieldRequest": {
            "type": "object",
            "required": [
                "location",
                "name",
                "price_per_hour"
            ],
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_per_hour": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "take-home-test_app_models.FieldResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_per_hour": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "take-home-test_app_models.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
...
        "take-home-test_app_models.UpdateFieldRequest": {
            "type": "object",
            "required": [
                "location",
                "name",
                "price_per_hour"
            ],
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price_per_hour": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}